причесывание комментов
