+++ yamlParser.py
@@ -1,0 +1,23 @@
+import yaml
+
+
+def getDataContract():
+    """
+    Функция считывает yaml
+    :return: все данные контракта в объекте
+    """
+    with open(r'C:\Users\Алексей\Documents\example_yaml\FinancialInstrument v1.15.yaml', 'rt', encoding='utf-8') as ya:
+        contract = ya.read()
+    data = yaml.safe_load(contract)
+    return data
+
+# df = pd.DataFrame(data) export_excel = df.to_excel(
+# r'C:\Users\Алексей\PycharmProjects\YamlContractParser\export_dataframe.xlsx', index=None, header=False)
+
+# TODO: добавить вывод в excel построчно
+
+# parametrs = Params(getDataContract()['basePath'], path=subPaths, methodName=methodNames,
+#                    paramName=pamNames, paramDiscrip=paramDescrip, paramType=pamTypes)
+# print(parametrs.subPath + parametrs.path[1] + '/' + parametrs.paramName[1])
+
+# todo:консольный интерфейс: colorama - цветные сообщения, pyinstaller - упаковщик в exe
